ATM and Age calculate
project-1 Survival Duration Calculator Project 

The objective of the project is to calculate the duration of how long a person lived based on his Age. The duration should be calculated in different time units like Months, Weeks, Days, Hours, Minutes, Seconds
The input to the application is an Age of a person and the output is expected to calculate the duration of time he lived in different time units as selected by the user.

project 2- ATM

the objective the project is to built an application which  is exactly working as a atm machine .you need  the include basic functionality such  as check balance ,withdraw,deposit  and exit  of the atm machine.